pyproject.toml
[project]
name = "object-detection"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "accelerate>=1.10.1",
    "gradio>=4.0.0",
    "moviepy>=2.2.1",
    "natsort>=8.4.0",
    "opencv-python>=4.12.0.88",
    "pillow>=8.0,<11.0",
    "six>=1.17.0",
    "torch>=2.8.0",
    "transformers>=4.57.1",
]
